refactor(habitat): simplify D2 registry boundary
