+++ tools/habitat-harness/src/rules/registry/load.ts
@@ -1,0 +1,102 @@
+import { readFileSync } from "node:fs";
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+import { Value } from "typebox/value";
+import {
+  RuleRegistryDocumentV1Schema,
+  type RuleRegistryDocumentV1,
+  type RuleRegistryRecordV1,
+} from "./schema.js";
+
+export type RuleRegistryIssueCode =
+  | "registry-json-invalid"
+  | "registry-schema-invalid"
+  | "registry-duplicate-rule-id";
+
+export interface RuleRegistryIssue {
+  code: RuleRegistryIssueCode;
+  path: string;
+  message: string;
+}
+
+export type RuleRegistryParseResult =
+  | { ok: true; document: RuleRegistryDocumentV1 }
+  | { ok: false; issues: RuleRegistryIssue[] };
+
+export function parseRuleRegistryText(
+  text: string,
+  sourcePath = "rules.json"
+): RuleRegistryParseResult {
+  let parsed: unknown;
+  try {
+    parsed = JSON.parse(text) as unknown;
+  } catch (error) {
+    return {
+      ok: false,
+      issues: [
+        {
+          code: "registry-json-invalid",
+          path: sourcePath,
+          message: error instanceof Error ? error.message : "Invalid JSON.",
+        },
+      ],
+    };
+  }
+
+  return parseRuleRegistryDocument(parsed, sourcePath);
+}
+
+export function parseRuleRegistryDocument(
+  value: unknown,
+  sourcePath = "rules.json"
+): RuleRegistryParseResult {
+  const schemaIssues = [...Value.Errors(RuleRegistryDocumentV1Schema, value)].map((error) => ({
+    code: "registry-schema-invalid" as const,
+    path: error.instancePath ? `${sourcePath}${error.instancePath}` : sourcePath,
+    message: error.message,
+  }));
+  if (schemaIssues.length > 0) return { ok: false, issues: schemaIssues };
+
+  const document = value as RuleRegistryDocumentV1;
+  const duplicateIssues = duplicateRuleIdIssues(document.rules, sourcePath);
+  if (duplicateIssues.length > 0) return { ok: false, issues: duplicateIssues };
+
+  return { ok: true, document };
+}
+
+export function loadRuleRegistryDocument(
+  registryPath = defaultRuleRegistryPath()
+): RuleRegistryDocumentV1 {
+  const result = parseRuleRegistryText(readFileSync(registryPath, "utf8"), registryPath);
+  if (!result.ok) {
+    throw new Error(
+      `Habitat rule registry is invalid:\n${result.issues
+        .map((issue) => `- ${issue.path}: ${issue.message}`)
+        .join("\n")}`
+    );
+  }
+  return result.document;
+}
+
+export function defaultRuleRegistryPath(): string {
+  return path.join(path.dirname(fileURLToPath(import.meta.url)), "..", "rules.json");
+}
+
+function duplicateRuleIdIssues(
+  rules: readonly RuleRegistryRecordV1[],
+  sourcePath: string
+): RuleRegistryIssue[] {
+  const seen = new Set<string>();
+  const duplicateIds = new Set<string>();
+  for (const rule of rules) {
+    if (seen.has(rule.id)) duplicateIds.add(rule.id);
+    seen.add(rule.id);
+  }
+  return [...duplicateIds].sort().map((id) => ({
+    code: "registry-duplicate-rule-id",
+    path: sourcePath,
+    message: `Duplicate Habitat rule id: ${JSON.stringify(id)}.`,
+  }));
+}
+
+export const activeRuleRegistryDocument = loadRuleRegistryDocument();
